refactor: 기존 코드를 MVC 패턴으로 리팩토링
- 패키지를 controller, model, view로 나누며 코드 역할 분리
- 코드 간소화 및 가독성 향상
